+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Compatibility:
 * Assigning to a numbered parameter raises `SyntaxError` (#2506, @andrykonchin).
 * Implement `--backtrace-limit` option (#2453).
 * Update `String` methods to return `String` instances when called on a subclass (#2453).
+* Relax `Fiber#transfer` limitation (#2453).
 
 Performance: